usrp2: Include DSP bit gain derived from 'peak' stream arg.
This enables use of the 'peak' stream arg in sc16 and sc8
over-the-wire modes to specify the peak value relative to 1.0
expected at the ADC input. Any resulting positive scaling required
in the DSP (from peak < 1.0) is then translated into a bit shift
and a scale factor applied in the FPGA. Using a bit shift allows
for precision gains due to decimation to be recovered by
sacrificing high order bits.